requirements.txt
boto3==1.34.3
langchain==0.0.340
opensearch-py==2.4.1
requests-aws4auth==1.2.3
python-dotenv